requirements.txt
# Core LangChain and LangGraph
langchain>=0.1.0,<0.3.0
langchain-community>=0.0.20
langgraph>=0.0.40
langchain-core>=0.1.0

# LLM Provider
groq>=0.4.0

# Vector Store and Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# Document Processing
pypdf>=3.0.0
python-dotenv>=1.0.0

# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Web Interface
streamlit>=1.28.0
plotly>=5.15.0

# Evaluation (optional - will be handled gracefully if missing)
rouge-score>=0.1.2
bert-score>=0.3.13

# Utilities
pydantic>=2.0.0
typing-extensions>=4.8.0
requests>=2.28.0

# Optional Tracing (install separately if needed)
# langsmith>=0.1.0
# trulens-eval>=0.20.0